pyproject.toml
[project]
name = "iphone-lockscreen-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.40",
    "icalendar>=6.3.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pytest>=8.4.2",
    "pytz>=2025.2",
    "recurring-ical-events>=3.8.0",
    "requests>=2.32.5",
    "toml>=0.10.2",
]

[dependency-groups]
dev = [
    "moto>=5.1.13",
]
